Introduce new modules "sort.js" for components. Use one for GScan based on the existing one from Tree. New elements added to GScan structure are added in order now.
